[FIX] FIx flake8
