+++ product_customer_code/views/product_product_view.xml
@@ -2,22 +2,20 @@
 <openerp>
     <data>
         
-        <record model="ir.ui.view" id="product_template_form_inh_product_customer_code_01">
-            <field name="name">product.template.form.inh.product.customer.code.01</field>
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="product.product_template_form_view"/>
+        <record model="ir.ui.view" id="product_normal_form_inh_product_customer_code_01">
+            <field name="name">product.normal.form.inh.product.customer.code.01</field>
+            <field name="model">product.product</field>
+            <field name="inherit_id" ref="product.product_normal_form_view"/>
             <field name="arch" type="xml">
                 <page string="Information" position="after">
                     <page string="Customer Code">
-                         <field name="product_customer_code_ids" nolabel="1" colspan="4" >
+                         <field name="product_customer_code_ids" nolabel="1" colspan="4" widget="one2many_list">
                             <tree string="Product Customer Code" editable="bottom">
                                 <field name="product_code"/>
-                                <field name="product_name"/>
                                 <field name="partner_id"/>
                             </tree>
-                            <form string="Product Customer Code">
+                            <form string="Product Customer Code" version="7.0">
                                 <field name="product_code"/>
-                                <field name="product_name"/>
                                 <field name="partner_id"/>
                             </form> 
                         </field>
